backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis for caching/rate limiting
redis==5.0.1

# GIS utilities
shapely==2.0.2
geojson==3.1.0

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
